+++ Ambassador/src/utopia/ambassador/rest/resource/service/auth/AuthNode.scala
@@ -16,7 +16,7 @@ import utopia.exodus.rest.util.AuthorizedContext
 import utopia.exodus.util.ExodusContext.logger
 import utopia.exodus.util.ExodusContext.uuidGenerator
 import utopia.flow.collection.immutable.caching.iterable.LazyVector
-import utopia.flow.collection.template.MapLike
+import utopia.flow.collection.template.MapAccess
 import utopia.flow.time.Now
 import utopia.flow.collection.CollectionExtensions._
 import utopia.nexus.http.Path
@@ -30,7 +30,7 @@ import utopia.vault.database.Connection
   * @author Mikko Hilpinen
   * @since 18.7.2021, v1.0
   */
-class AuthNode(target: ServiceTarget, tokenAcquirer: AcquireTokens, redirectors: MapLike[Int, AuthRedirector])
+class AuthNode(target: ServiceTarget, tokenAcquirer: AcquireTokens, redirectors: MapAccess[Int, AuthRedirector])
 	extends ResourceWithChildren[AuthorizedContext]
 {
 	// ATTRIBUTES   -----------------------------