+++ Flow/Changes.md
@@ -22,6 +22,7 @@
 - Renamed **Zeroable** to **CanBeZero** and **ApproximatelyZeroable** to **CanBeAboutZero**
   - **CanBeZero** now also requires an implementation for a new abstract property `zero: Repr`
 - Renamed **LinearMeasurable** to **HasLength**
+- Renamed **MapLike** to **MapAccess**
 - **TreeLike**, **GraphNode** and **GraphEdge** no longer extend **Node**
 - Renamed **ConversionReliability** values to PascalCase (e.g. from **NO_CONVERSION** to **NoConversion**)
 ### Deprecations